+++ src/main/java/BobException.java
@@ -1,0 +1,5 @@
+public class BobException extends Exception {
+    public BobException(String msg) {
+        super(msg);
+    }
+}
