+++ src/main/java/Deadline.java
@@ -1,6 +1,6 @@
 public class Deadline extends Task {
     private String end;
-    public Deadline(String description, String end) throws EchoMIndException {
+    public Deadline(String description, String end) throws BobException {
         super(description);
         this.end = end;
     }
