--- src/main/java/EchoMIndException.java
@@ -1,5 +1,0 @@
-public class EchoMIndException extends Exception {
-    public EchoMIndException(String msg) {
-        super(msg);
-    }
-}
