+++ src/main/java/Event.java
@@ -1,7 +1,7 @@
 public class Event extends Task {
     private String start;
     private String end;
-    public Event(String description, String start, String end) throws EchoMIndException {
+    public Event(String description, String start, String end) throws BobException {
         super(description);
         this.start = start;
         this.end = end;
