+++ text-ui-test/input.txt
@@ -7,4 +7,10 @@ mark 3
 list
 unmark 2
 list
+todo
+deadline
+event
+delete 2
+list
+blahblahblah
 bye
